requirement.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
distlib==0.3.6
Django==4.1.4
django-recaptcha==3.0.0
django-session-timeout==0.1.0
exceptiongroup==1.1.0
filelock==3.9.0
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h11==0.14.0
h5py==3.7.0
idna==3.4
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==22.0
platformdirs==2.6.2
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dotenv==0.21.0
python-http-client==3.3.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
selenium==4.7.2
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.17.1
webdriver-manager==3.8.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xlwt==1.3.0
